Fix rostest target names when build dir is inside source dir (#2361)
It's a common approach to build cmake projects in a `build` folder in the project root.
It's just that `rosbuild`/`catkin` does not do it, so the changed name never mattered.

When you try the target name includes another `build_` string and there is *exactly one* package in the whole ecosystem that relies on this not happening:
https://github.com/ros/ros_comm/blob/845f74602c7464e08ef5ac6fd9e26c97d0fe42c9/test/test_rosbag/bag_migration_tests/CMakeLists.txt#L87-L103

(cherry picked from commit 336b29ab103e8c496e495215abd49df89c7a218d)